==> Scripts/Script2_regroupement_taxonomique_300m.Rmd <==
---
title: "Script2_regroupement_taxonomique_300m"
date: "4/12/2025"
author: "Julie Dutoict (julie.dutoict@etu.sorbonne-universite.fr)"
output: html_document
---

 Ce script associe à chaque taxon selectionné sous Ecotaxa un taxon choisi par l'utilisateur : 

## 1. Chargement des packages et des données 

```{r}
library('tidyverse')
```


Télécharger le tableau vierge créé dans script 1 et récupérer le jeu de données Ecotaxa300m

```{r}
taxa_rough <- read.table("../Output/Intermediary Data/table_taxons.tsv", sep = ",", header = TRUE)
ecotaxa300m <- read_csv("../Output/Intermediary Data/Ecotaxa300m.csv")
colnames(ecotaxa300m)[colnames(ecotaxa300m) == "object_annotation_category"] <- "taxon"
```

## 2. Association des nouveaux taxons choisis aux anciens 

Trier dans l'ordre alphabétique et associer un nouveau taxon à chaque catégorie taxonomique

```{r}
taxa_rough$taxon <- sort(taxa_rough$taxon)
taxa_rough$taxo_rough <- c("Na",
                           "Annelida",
                           "Autres Crustaces",
                           "Annelida",
                           "Appendicularia",
                           "Na",
                           "Phaeodaria",
                           "Phaeodaria",
                           "Phaeodaria",
                           "Copepoda",
                           "Bean-like",
                           "Copepoda",
                           "Copepoda",
                           "Mollusca",
                           "Mollusca",
                           "Chaetognatha",
                           "Na",
                           "Mollusca",
                           "Autres Collodaria",
                           "Phaeodaria",
                           "Phaeodaria",
                           "Autres Collodaria",
                           "Autres Collodaria",
                           "Autres Rhizaria",
                           "Copepoda",
                           "Copepoda",
                           "Autres Crustaces",
                           "Na",
                           "Na",
                           "Na",
                           "Appendicularia",
                           "Detritus",
                           "Na",
                           "Copepoda",
                           "Autres Crustaces",
                           "Autres Crustaces",
                           "Detritus",
                           "Detritus",
                           "Autres Rhizaria",
                           "Mollusca",
                           "Hydrozoa",
                           "Autres Crustaces",
                           "Copepoda-like",
                           "Mollusca",
                           "Phaeodaria",
                           "Na",
                           "Hydrozoa",
                           "Ostracoda",
                           "Na",
                           "Na",
                           "Hydrozoa",
                           "Autres Crustaces",
                           "Phaeodaria",
                           "Annelida",
                           "Annelida",
                           "Na",
                           "Na",
                           "Na",
                           "Autres Rhizaria",
                           "Na",
                           "Na",
                           "Autres Crustaces",
                           "Hydrozoa",
                           "Solitary black",
                           "Solitary globule",
                           "Hydrozoa",
                           "Autres Rhizaria",
                           "Phaeodaria",
                           "Phaeodaria",
                           "Autres Rhizaria",
                           "Annelida",
                           "Na",
                           "Na",
                           "Na",
                           "Na",
                           "Na",
                           "Hydrozoa",
                           "Mollusca",
                           "Annelida",
                           "Hydrozoa"
)
```

Créer une colonne "taxo_rough" avec les nouveaux taxons dans le tableau Ecotaxa300m

```{r}
map_taxo <- setNames(taxa_rough$taxo_rough, taxa_rough$taxon)
ecotaxa300m <- ecotaxa300m %>%
  mutate(taxo_rough = map_taxo[taxon])
names(ecotaxa300m)[ncol(ecotaxa300m)]
```

## 3. Vérification des correspondances 

On vérifie qu'il n'y a pas de catégories sans correspondance taxonomique (taxons orphelins)

```{r}
orphans <- ecotaxa300m %>%
  filter(is.na(taxo_rough)) %>%
  distinct(taxon)

cat("Nombre de taxons orphelins :", nrow(orphans), "\n")
```

## 4. Comptage des objects par nouveau groupe taxonomique

```{r}
count_taxo_rough <- ecotaxa300m %>%
  group_by(taxo_rough) %>%
  summarise(n = n(), .groups = "drop") %>%
  filter(!is.na(taxo_rough))
```

On regarde combien on a d'objets en tout (vérification : 841 747)

```{r}
cat("Nombre total d’images regroupées:", sum(count_taxo_rough$n), "\n")
```

On renomme les colonnes

```{r}
colnames(count_taxo_rough)[colnames(count_taxo_rough) == "taxo_rough"] <- "New taxons"
colnames(count_taxo_rough)[colnames(count_taxo_rough) == "n"] <- "Count"
```

## 5. Sauvegarde des données

Sauvegarder le tableau des abondances de chaque taxon pour toute la série temporelle

```{r}
write_csv(count_taxo_rough, "../Output/Final Data/Count_total_taxo_300m.csv")
write_csv(ecotaxa300m, "../Output/Intermediary Data/Ecotaxa300m_taxons.csv")
```

vérification finale : le nombre de lignes du le jeu de données (= nombre d'objets) doit être égal à la somme des abondances de la série temporelle

```{r}
nrow(ecotaxa300m) == sum(count_taxo_rough$Count, na.rm = TRUE)
```
