==> Scripts/Script3_groupement_abondance_mois_300m.Rmd <==
---
title: "Script3_groupement_abondance_mois_300m"
date: "4/12/2025"
author: "Julie Dutoict (julie.dutoict@etu.sorbonne-universite.fr)"
output: html_document
---

Ce script groupe les abondances par mois et reformatte le tableau pour obtenir un tableau avec colonnes = taxon et ligne = date

## Setup

```{r}
library(dplyr)
library(tidyr)
library(lubridate)
library(readr)

ecotaxa300m <- read.table("../Output/Intermediary Data/Ecotaxa300m_taxons.csv", sep = ",", header = TRUE)

```

## Création du tableau journalier

Enlever les copépodes-like prédits car ils sont mal identifiés,

```{r}
ecotaxa300m_NOCOP <- ecotaxa300m %>%
  filter(!(object_annotation_status == "predicted" & taxon == "Copepoda-like"))
```

Séquence de toutes les dates attendues

```{r}
dates_attendues <- seq.Date(from = as.Date("2019-10-12"), 
                            to = as.Date("2025-02-01"), 
                            by = "day")
```

Groupement des abondances par date et taxon

```{r}
eco_300_summary <- ecotaxa300m_NOCOP %>%
  select(Date = object_date, Heure = object_time, Taxa = taxo_rough) %>%
  mutate(Date = as.Date(Date))  # Colonne bien au format Date

tableau_jour <- eco_300_summary %>%
    group_by(Date, Taxa) %>%
    summarise(abondance = n(), .groups = "drop") %>%
    ## Compléter toutes les dates + tous les taxons
    complete(Date = dates_attendues, Taxa, fill = list(abondance = 0)) %>%
    ## Pivotage
    pivot_wider(
      names_from = Taxa,
      values_from = abondance,
      values_fill = list(abondance = 0)
  )
```

Enlever la colonne Na et vérifier le nombre de dates

```{r}
tableau_jour$Na <- NULL
nrow(tableau_jour) # = 1940 perfect
```

Rajouter les colonnes total contage + total sans détritus

```{r}
tableau_jour <- tableau_jour %>%
  mutate(
    Total = rowSums(select(., -Date)),
    `Total Zooplancton` = rowSums(select(., -Date, -Detritus))
  )
```

## Création du tableau mensuel

Convertir la colonne Date (si cest pas déjà fait)

```{r}
tableau_jour$Date <- as.Date(tableau_jour$Date)
```

Ajouter une colonne "mois"

```{r}
tableau_jour$Month <- format(tableau_jour$Date, "%Y-%m")
```

Grouper par mois et sommer toutes les autres colonnes sauf Date

```{r}
table_month <- tableau_jour %>%
  select(-Date) %>%
  group_by(Month) %>%
  summarise(across(everything(), \(x) sum(x, na.rm = TRUE)))
```

Télécharger le tableau

```{r}
write_csv(table_month, "../Output/Intermediary Data/Table_month_zoo_NOCOP_300m.csv")
```
